Change time precision to double throughout code

+++ src/kbmod/search/image_stack.h
@@ -26,8 +26,8 @@ public:
 
     // Functions for getting times.
     double get_obstime(int index) const;
-    float get_zeroed_time(int index) const;
-    std::vector<float> build_zeroed_times() const;  // Linear cost.
+    double get_zeroed_time(int index) const;
+    std::vector<double> build_zeroed_times() const;  // Linear cost.
 
     void convolve_psf();
