+++ src/kbmod/search/psi_phi_array_utils.h
@@ -28,7 +28,7 @@ namespace search {
 std::array<float, 3> compute_scale_params_from_image_vect(const std::vector<RawImage>& imgs, int num_bytes);
 
 void fill_psi_phi_array(PsiPhiArray& result_data, int num_bytes, const std::vector<RawImage>& psi_imgs,
-                        const std::vector<RawImage>& phi_imgs, const std::vector<float> zeroed_times,
+                        const std::vector<RawImage>& phi_imgs, const std::vector<double> zeroed_times,
                         bool debug = false);
 
 void fill_psi_phi_array_from_image_stack(PsiPhiArray& result_data, ImageStack& stack, int num_bytes,
